% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{baseEstimate}
\alias{baseEstimate}
\title{Estimating Projection Matrices}
\usage{
baseEstimate(size, magn, Y, X1, X2, Kernlist, mode, lambda)
}
\arguments{
\item{size}{A numeric number specifying the number of observations.}

\item{magn}{A numeric number specifying the number of kernels in the kernel
library.}

\item{Y}{Reponses of the dataframe.}

\item{X1}{The first type of factor in the dataframe (could contains several
subfactors).}

\item{X2}{The second type of factor in the dataframe (could contains several
subfactors).}

\item{Kernlist}{The kernel library containing several kernels given by user.}

\item{mode}{A character string indicating which tuning parameter criteria is
to be used.}

\item{lambda}{A numeric string specifying the range of noise to be chosen.
The lower limit of lambda must be above 0.}
}
\value{
\item{A_hat}{A list of projection matrices for every kernels in the
kernel library.}

\item{error_mat}{A size*magn matrix indicating errors.}
}
\description{
Calculate the estiamted projection matrices for every kernels in the kernel
library.
}
\details{
For a given mode, this function return a list of projection matrices for
every kernels in the kernel library and a size*magn matrix indicating
errors.
}
\examples{


##baseEstimate(size = 50, magn = 3, Y, X1, X2, Kernlist = NULL,
##mode = "loocv", lambda = exp(seq(-5, 5)))


}
\references{
Jeremiah Zhe Liu and Brent Coull. Robust Hypothesis Test for
Nonlinear Effect with Gaus- sian Processes. October 2017.
}
\author{
Wenying Deng
}
