% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{compute_stat}
\alias{compute_stat}
\title{Computing Score Test Statistics.}
\usage{
compute_stat(n, Y, Z12, y_fixed, K_gpr, sigma2_hat, tau_hat)
}
\arguments{
\item{n}{(integer) A numeric number specifying the number of observations.}

\item{Y}{(vector of length n) Reponses of the dataframe.}

\item{Z12}{(dataframe, n*(q1\*q2)) The interaction items of the first and
second groups of random effects variables in the dataframe.}

\item{y_fixed}{(vector of length n) Estimated fixed effects of the
responses.}

\item{K_gpr}{(matrix, n*n) Estimated ensemble kernel matrix.}

\item{sigma2_hat}{(numeric) The estimated noise of the fixed effects.}

\item{tau_hat}{(numeric) The estimated noise of the random effects.}
}
\value{
\item{test_stat}{(numeric) The computed test statistic.}
}
\description{
Compute score test statistics.
}
\details{
The test statistic is distributed as a scaled Chi-squared distribution.
}
\examples{



compute_stat(n = 100, Y, Z12, y_fixed, K_ens, sigma2_hat, tau_hat)



}
\references{
Arnab Maity and Xihong Lin. Powerful tests for detecting a gene
effect in the presence of possible gene-gene interactions using garrote
kernel machines. December 2011.
}
\author{
Wenying Deng
}
