% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{cvek_test}
\alias{cvek_test}
\title{Conduct Hypothesis Testing based on CVEK estimation result.}
\usage{
cvek_test(est_res, formula_test, kern_func_list, data, test = "boot",
  alt_kernel_type = "linear", B = 100, verbose = FALSE)
}
\arguments{
\item{est_res}{(list) Estimation results returned by estimation() procedure.}

\item{formula_test}{(formula) A user-supplied formula indicating the alternative
effect to test. All terms in the alternative mode must be specified as kernel terms.}

\item{kern_func_list}{(list) a list of kernel functions in the kernel library}

\item{data}{(data.frame, n*d) a data.frame, list or environment (or object
coercible by as.data.frame to a data.frame), containing the variables in
formula. Neither a matrix nor an array will be accepted.}

\item{test}{(character) Type of hypothesis test to conduct.
Must be eitehr 'asymp' or 'boot'.}

\item{alt_kernel_type}{(character) Type of alternative kernel effect to consider.
Must be either 'linear' or 'ensemble'}

\item{B}{(integer) A numeric value indicating times of resampling when test
= "boot".}

\item{verbose}{(logical) Whether to print additional messages.}
}
\value{
\item{pvalue}{(numeric) p-value of the test.}
}
\description{
Conduct Hypothesis Testing based on CVEK estimation result.
}
\keyword{internal}
