% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{dataGenerate}
\alias{dataGenerate}
\title{Generating Original Data}
\usage{
dataGenerate(size, label_names, method = "rbf", int_effect = 0, l = 1,
  p = 2, eps = 0.01)
}
\arguments{
\item{size}{A numeric number specifying the number of observations.}

\item{label_names}{A character string indicating all the interior variables
included in each predictor.}

\item{method}{A character string indicating which kernel is to be computed.}

\item{int_effect}{A numeric number specifying the size of interaction.}

\item{l}{A numeric number indicating the hyperparameter (flexibility) of a
specific kernel.}

\item{p}{For polynomial, p is the power; for matern, v = p + 1 / 2; for
rational, alpha = p.}

\item{eps}{A numeric number indicating the size of noise.}
}
\value{
\item{data}{A dataframe to be fitted.}
}
\description{
Generate original data based on specific kernels.
}
\details{
This function generates with a specific kernel. The argument int_effect
represents the strength of interaction relative to the main effect since all
sampled functions have been standardized to have unit norm.
}
\examples{


##data <- dataGenerate(size = 100, label_names =
##list(X1 = c("x1", "x2"), X2 = c("x3", "x4")),
##method = "rbf", int_effect = 0, l = 1, p = 2, eps = 0.01)


}
\author{
Wenying Deng
}
