% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineModel.R
\name{defineModel}
\alias{defineModel}
\title{Defining the Model}
\usage{
defineModel(formula, label_names, data, Kern_par)
}
\arguments{
\item{formula}{A symbolic description of the model to be fitted.}

\item{label_names}{A character string indicating all the interior variables
included in each predictor.}

\item{data}{A dataframe to be fitted.}

\item{Kern_par}{A dataframe indicating the parameters of base kernels to be
created.}
}
\value{
\item{Y}{Reponses of the dataframe.}

\item{X1}{The first type of factor in the dataframe (could contains several
subfactors).}

\item{X2}{The second type of factor in the dataframe (could contains several
subfactors).}

\item{Kernlist}{The kernel library containing several kernels given by
user.}
}
\description{
Give the complete formula and generate the expected kernel library.
}
\details{
It processes data based on formula and label_names and creates a kernel
library according to the parameters given in Kern_par.
}
\examples{


##Kern_par <- data.frame(method = c("rbf", "polynomial", "matern"),
##Sigma = rep(0, 3), l = c(.5, 1, 1.5), p = 1:3)
##Kern_par$method <- as.character(Kern_par$method)
##defineModel(formula = Y ~ X1 + X2,
##label_names = list(X1 = c("x1", "x2"), X2 = c("x3", "x4")),
##data, Kern_par)


}
\seealso{
method: \code{\link{kernelGenerate}}
}
\author{
Wenying Deng
}
