% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dora.R
\docType{data}
\name{dora}
\alias{dora}
\title{Dataset as an example}
\format{A data frame with 100 rows and 5 variables}
\usage{
data(dora)
}
\description{
A dataset containing a response variable and four predictor variables. 
The variables are as follows:
}
\details{
\itemize{
  \item y. (vector of length n) Reponses of the dataframe.
  \item x1. (dataframe, n*1) The first variable in the dataframe.
  \item x2. (dataframe, n*1) The second variable in the dataframe.
  \item x3. (dataframe, n*1) The third variable in the dataframe.
  \item x4. (dataframe, n*1) The fourth variable in the dataframe.
}
}
\keyword{datasets}
