% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{ensemble_kernel_matrix}
\alias{ensemble_kernel_matrix}
\title{Calculating Ensemble Kernel Matrix}
\usage{
ensemble_kernel_matrix(A_est, eig_thres = 1e-11)
}
\arguments{
\item{A_est}{(matrix) Ensemble projection matrix.}

\item{eig_thres}{(numeric) Threshold to truncate the kernel matrix.}
}
\value{
\item{K_hat}{(matrix, n*n) Estimated ensemble kernel matrix.}
}
\description{
Calculating ensemble kernel matrix and truncating those columns whose
eigenvalues are smaller than the given threshold.
}
\details{
After we obtain the ensemble projection matrix, we can calculate the
ensemble kernel matrix.
}
\examples{



ensemble_kernel_matrix(A_hat[[2]], eig_thres = 1e-11)



}
\author{
Wenying Deng
}
