% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{estimate_base}
\alias{estimate_base}
\title{Estimating Projection Matrices}
\usage{
estimate_base(Y, X, K_list, mode, lambda)
}
\arguments{
\item{Y}{(vector of length n) Reponses of the dataframe.}

\item{X}{(dataframe, n*p) Fixed effects variables in the dataframe (could
contains several subfactors).}

\item{K_list}{(list of matrices) A nested list of kernel term matrices. 
The first level corresponds to each base kernel function in kern_func_list, 
the second level corresponds to each kernel term specified in the formula.}

\item{mode}{(character) A character string indicating which tuning parameter
criteria is to be used.}

\item{lambda}{(numeric) A numeric string specifying the range of noise to be
chosen. The lower limit of lambda must be above 0.}
}
\value{
\item{A_hat}{(list of length K) A list of projection matrices for
each kernel in the kernel library.}

\item{P_K_hat}{(list of length K) A list of projection matrices to kernel
space for each kernel in the kernel library.}

\item{beta_list}{(list of length K) A list of fixed effects estimator for
each kernel in the kernel library.}

\item{alpha_list}{(list of length K) A list of random effects estimator for
each kernel in the kernel library.}

\item{lambda}{(list of length K) A list of selected tuning parameters for
each kernel in the kernel library.}

\item{error_mat}{(matrix, n*K) A n\*K matrix indicating errors.}
}
\description{
Calculate the estiamted projection matrices for every kernels in the kernel
library.
}
\details{
For a given mode, this function return a list of projection matrices for
every kernels in the kernel library and a n*K matrix indicating errors.
}
\references{
Jeremiah Zhe Liu and Brent Coull. Robust Hypothesis Test for
Nonlinear Effect with Gaus- sian Processes. October 2017.
}
\author{
Wenying Deng
}
\keyword{internal}
