% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{estimate_noise}
\alias{estimate_noise}
\title{Estimating Noise}
\usage{
estimate_noise(Y, lambda_hat, beta_hat, alpha_hat, K_hat)
}
\arguments{
\item{Y}{(vector of length n) Reponses of the dataframe.}

\item{lambda_hat}{(numeric) The selected tuning parameter based on the 
estimated ensemble kernel matrix.}

\item{beta_hat}{(numeric) Estimated bias of the model.}

\item{alpha_hat}{(vector of length n) Estimated coefficients of the estimated 
ensemble kernel matrix.}

\item{K_hat}{(matrix, n*n) Estimated ensemble kernel matrix.}
}
\value{
\item{sigma2_hat}{(numeric) The estimated noise of the fixed effects.}
}
\description{
An implementation of Gaussian processes for estimating noise.
}
\examples{


sigma2_hat <- estimate_noise(Y, lam, beta0, alpha0, K_gpr)


}
\references{
Jeremiah Zhe Liu and Brent Coull. Robust Hypothesis Test for
Nonlinear Effect with Gaus- sian Processes. October 2017.
}
\author{
Wenying Deng
}
