% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{euc_dist}
\alias{euc_dist}
\title{Computing Euclidean Distance between Two Vectors (Matrices)}
\usage{
euc_dist(x1, x2 = NULL)
}
\arguments{
\item{x1}{(vector/matrix) The first vector/matrix.}

\item{x2}{(vector/matrix, the same dimension as x1) 
The second vector/matrix.}
}
\value{
\item{dist}{(numeric) Euclidean distance.}
}
\description{
Compute the L2 distance between two vectors or matrices.
}
\details{
This function gives the Euclidean distance between two 
vectors or matrices.
}
\author{
Wenying Deng
}
\keyword{internal}
