% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{generate_data}
\alias{generate_data}
\title{Generating Original Data}
\usage{
generate_data(n, label_names, method = "rbf", int_effect = 0, l = 1,
  p = 2, eps = 0.01)
}
\arguments{
\item{n}{(integer) A numeric number specifying the number of observations.}

\item{label_names}{(list) A character string indicating all the interior 
variables included in each predictor.}

\item{method}{(character) A character string indicating which kernel is 
to be computed.}

\item{int_effect}{(numeric) A numeric number specifying the size of interaction.}

\item{l}{(numeric) A numeric number indicating the hyperparameter 
(flexibility) of a specific kernel.}

\item{p}{(integer) For polynomial, p is the power; for matern, v = p + 1 / 2; 
for rational, alpha = p.}

\item{eps}{(numeric) A numeric number indicating the size of noise.}
}
\value{
\item{data, n*P}{(dataframe) A dataframe to be fitted.}
}
\description{
Generate original data based on specific kernels.
}
\details{
This function generates with a specific kernel. The argument int_effect
represents the strength of interaction relative to the main effect since all
sampled functions have been standardized to have unit norm.
}
\examples{


data <- generate_data(n = 100, label_names =
list(X1 = c("x1", "x2"), X2 = c("x3", "x4")),
method = "rbf", int_effect = 0, l = 1, p = 2, eps = .01)


}
\author{
Wenying Deng
}
