% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{genericFormula}
\alias{genericFormula}
\title{From Vectors to Single Variables}
\usage{
genericFormula(formula, label_names)
}
\arguments{
\item{formula}{A symbolic description of the model to be fitted.}

\item{label_names}{A character string indicating all the interior variables
included in each predictor.}
}
\value{
\item{generic_formula}{A symbolic description of the model written
in single variables format.}

\item{length_main}{A numeric value indicating the length of main effects.}
}
\description{
Transform format of predictors from vectors to single variables.
}
\examples{


##generic_formula0 <- genericFormula(formula = Y ~ X1 + X2,
##label_names = list(X1 = c("x1", "x2"), X2 = c("x3", "x4")))


}
\author{
Wenying Deng
}
