% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{infoMat}
\alias{infoMat}
\title{Computing Information Matrices}
\usage{
infoMat(P0_mat, mat_del = NULL, mat_sigma2 = NULL, mat_tau = NULL)
}
\arguments{
\item{P0_mat}{Scale projection matrix under REML.}

\item{mat_del}{Derivative of the scale covariance matrix of Y with respect
to delta.}

\item{mat_sigma2}{Derivative of the scale covariance matrix of Y with
respect to sigma2.}

\item{mat_tau}{Derivative of the scale covariance matrix of Y with respect
to tau}
}
\value{
\item{I0}{The computed information value.}
}
\description{
Compute information matrices based on block matrices.
}
\details{
This function gives the information value of the interaction strength.
}
\examples{


##I0 <- infoMat(P0_mat, mat_del = drV0_del,
##mat_sigma2 = drV0_sigma2, mat_tau = drV0_tau)


}
\references{
Arnab Maity and Xihong Lin. Powerful tests for detecting a gene
effect in the presence of possible gene-gene interactions using garrote
kernel machines. December 2011.
}
\author{
Wenying Deng
}
