% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kern_par.R
\docType{data}
\name{kern_par}
\alias{kern_par}
\title{Dataframe as an example for kernel parameters}
\format{A data frame with 3 rows and 3 variables}
\usage{
data(kern_par)
}
\description{
A dataframe containing the information from user to define base kernels.
}
\details{
\itemize{
  \item method. (character) A character string indicating which kernel 
is to be computed.
  \item l. (numeric) A numeric number indicating the hyperparameter 
(flexibility) of a specific kernel.
  \item d. (integer) For polynomial, d is the power; for matern, v = d + 1 / 2; for
rational, alpha = d.
}
}
\keyword{datasets}
