% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_kernel.R
\name{kernel_linear}
\alias{kernel_linear}
\title{Generating A Single vector-wise Function Using Linear}
\usage{
kernel_linear(l, p)
}
\arguments{
\item{l}{(numeric) A numeric number indicating the hyperparameter 
(flexibility) of a specific kernel.}

\item{p}{(integer) For polynomial, p is the power; for matern, v = p + 1 / 2; for
rational, alpha = p.}
}
\value{
\item{vector_wise}{(function) A function calculating 
the relevance of two vectors.}
}
\description{
Generate vector-wise functions for two vectors using linear kernel.
}
\details{
\bold{Polynomial Kernels} \deqn{k(x, x')=(x \cdot x')^p} We have intercept
kernel when \eqn{p=0}, and linear kernel when \eqn{p=1}.
}
\references{
The MIT Press. Gaussian Processes for Machine Learning, 2006.
}
\author{
Wenying Deng
}
