% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_kernel.R
\name{kernel_polynomial}
\alias{kernel_polynomial}
\title{Generating A Single Point-wise Function Using Polynomial}
\usage{
kernel_polynomial(Sigma, l, p)
}
\arguments{
\item{Sigma}{(matrix) The covariance matrix for neural network kernel.}

\item{l}{(numeric) A numeric number indicating the hyperparameter 
(flexibility) of a specific kernel.}

\item{p}{(integer) For polynomial, p is the power; for matern, v = p + 1 / 2; for
rational, alpha = p.}
}
\value{
\item{point_wise}{(function) A function calculating 
the relevance of two vectors.}
}
\description{
Generate point-wise functions for two vectors using polynomial kernel.
}
\details{
\bold{Polynomial Kernels} \deqn{k(x, x')=(x \cdot x')^p} We have intercept
kernel when \eqn{p=0}, and linear kernel when \eqn{p=1}.
}
\references{
The MIT Press. Gaussian Processes for Machine Learning, 2006.
}
\author{
Wenying Deng
}
