% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelTest.R
\name{kernelTest}
\alias{kernelTest}
\title{Conducting Score Tests for Interaction}
\usage{
kernelTest(formula, label_names, Kernlist, data = NULL, mode = "loocv",
  strategy = "erm", beta = 1, test = "boot", lambda = exp(seq(-5, 5)),
  B = 100)
}
\arguments{
\item{formula}{A symbolic description of the model to be fitted.}

\item{label_names}{A character string indicating all the interior variables
included in each predictor.}

\item{Kernlist}{The kernel library containing several kernels given by user.}

\item{data}{A dataframe to be fitted.}

\item{mode}{A character string indicating which tuning parameter criteria is
to be used.}

\item{strategy}{A character string indicating which ensemble strategy is to
be used.}

\item{beta}{A numeric value specifying the parameter when strategy = "exp".}

\item{test}{A character string indicating which test is to be used.}

\item{lambda}{A numeric string specifying the range of noise to be chosen.
The lower limit of lambda must be above 0.}

\item{B}{A numeric value indicating times of resampling when test = "boot".}
}
\value{
\item{pvalue}{p-value of the test.}
}
\description{
Conduct score tests comparing a fitted model and a more general alternative
model.
}
\details{
There are two tests available here:

\bold{Asymptotic Test}

This is based on the classical variance component test to construct a
testing procedure for the hypothesis about Gaussian process function.

\bold{Bootstrap Test}

When it comes to small sample size, we can use bootstrap test instead, which
can give valid tests with moderate sample sizes and requires similar
computational effort to a permutation test.
}
\examples{

##kernelTest(formula = Y ~ X1 + X2,
##label_names = list(X1 = c("x1", "x2"), X2 = c("x3", "x4")),
##Kernlist, data = data, mode = "loocv", strategy = "erm",
##beta = 1, test = "boot", lambda = exp(seq(-5, 5)))

}
\references{
Xihong Lin. Variance component testing in generalised linear
models with random effects. June 1997.

Arnab Maity and Xihong Lin. Powerful tests for detecting a gene effect in
the presence of possible gene-gene interactions using garrote kernel
machines. December 2011.

Petra Bu ̊zˇkova ́, Thomas Lumley, and Kenneth Rice. Permutation and
parametric bootstrap tests for gene-gene and gene-environment interactions.
January 2011.
}
\seealso{
method: \code{\link{kernelGenerate}}

mode: \code{\link{tuning}}

strategy: \code{\link{ensemble}}
}
\author{
Wenying Deng
}
