% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{multiScore}
\alias{multiScore}
\title{Computing Score Test Statistics.}
\usage{
multiScore(formula_int, label_names, Kernlist, data, mode, strategy, beta,
  lambda)
}
\arguments{
\item{formula_int}{A symbolic description of the model with interaction.}

\item{label_names}{A character string indicating all the interior variables
included in each predictor.}

\item{Kernlist}{The kernel library containing several kernels given by user.}

\item{data}{A dataframe to be fitted.}

\item{mode}{A character string indicating which tuning parameter criteria is
to be used.}

\item{strategy}{A character string indicating which ensemble strategy is to
be used.}

\item{beta}{A numeric value specifying the parameter when strategy = "exp".}

\item{lambda}{A numeric string specifying the range of noise to be chosen.
The lower limit of lambda must be above 0.}
}
\value{
\item{test_stat}{The computed test statistic.}
}
\description{
Compute score test statistics.
}
\details{
The test statistic is distributed as a scaled Chi-squared distribution.
}
\examples{

##multiScore(formula_int = Y ~ X1 + X2 + X1 * X2,
##label_names = list(X1 = c("x1", "x2"), X2 = c("x3", "x4")),
##Kernlist = Kernlist, data = data, mode = "loocv",
##strategy = "erm", test = "boot", 
##lambda = exp(seq(-5, 5)))

}
\references{
Arnab Maity and Xihong Lin. Powerful tests for detecting a gene
effect in the presence of possible gene-gene interactions using garrote
kernel machines. December 2011.
}
\author{
Wenying Deng
}
