% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{noiseEstimate}
\alias{noiseEstimate}
\title{Estimating Noise}
\usage{
noiseEstimate(Y, lambda_hat, beta_hat, alpha_hat, K_hat)
}
\arguments{
\item{Y}{Reponses of the dataframe.}

\item{lambda_hat}{The selected tuning parameter based on the estimated
ensemble kernel matrix.}

\item{beta_hat}{Estimated bias of the model.}

\item{alpha_hat}{Estimated coefficients of the estimated ensemble kernel
matrix.}

\item{K_hat}{Estimated ensemble kernel matrix.}
}
\value{
\item{sigma2_hat}{The estimated noise of the fixed effects.}
}
\description{
An implementation of Gaussian processes for estimating noise.
}
\examples{


##sigma2_hat <- noiseEstimate(y, lam, beta0, alpha0, K_gpr)


}
\references{
Jeremiah Zhe Liu and Brent Coull. Robust Hypothesis Test for
Nonlinear Effect with Gaus- sian Processes. October 2017.
}
\author{
Wenying Deng
}
