% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.cvek}
\alias{predict.cvek}
\title{Predicting New Response}
\usage{
\method{predict}{cvek}(object, data_new)
}
\arguments{
\item{object}{(list) Estimation results returned by estimation() procedure.}

\item{data_new}{(dataframe) The new set of predictors, whose name is 
the same as those of formula estimation().}
}
\value{
\item{y_pred}{(vector of length n) Predicted new response.}
}
\description{
Predicting new response based on given design matrix and 
the estimation result.
}
\details{
After we obtain the estimation result, we can predict new response.
}
\author{
Wenying Deng
}
