% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_kernel.R
\name{square_dist}
\alias{square_dist}
\title{Computing Square Distance between Two Sets of Variables}
\usage{
square_dist(X1, X2 = NULL, l = 1)
}
\arguments{
\item{X1}{(matrix, n1*p0) The first set of variables.}

\item{X2}{(matrix, n2*p0) The second set of variables.}

\item{l}{(numeric) A numeric number indicating the hyperparameter
(flexibility) of a specific kernel.}
}
\value{
\item{kern}{(function) A function calculating the relevance of two
matrices.}
}
\description{
Compute Squared Euclidean distance between two sets of variables with the
same dimension.
}
\references{
The MIT Press. Gaussian Processes for Machine Learning, 2006.
}
\author{
Wenying Deng
}
\keyword{internal}
