% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{standardize}
\alias{standardize}
\title{Standardizing Matrix}
\usage{
standardize(X)
}
\arguments{
\item{X}{(matrix, n*p0) Original data matrix.}
}
\value{
\item{X}{(matrix, n*p0) Standardized data matrix.}
}
\description{
Center and scale the data matrix into mean zero and standard deviation one.
}
\details{
This function gives the standardized data matrix.
}
\author{
Wenying Deng
}
\keyword{internal}
