% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{compute_stat}
\alias{compute_stat}
\title{Computing Score Test Statistics.}
\usage{
compute_stat(n, Y, X12, beta0, sigma2_hat, tau_hat, K_gpr)
}
\arguments{
\item{n}{(integer) A numeric number specifying the number of observations.}

\item{Y}{(vector of length n) Reponses of the dataframe.}

\item{X12}{(dataframe, n*(p1\*p2)) The interaction items of first and second 
types of factors in the dataframe.}

\item{beta0}{(numeric) Estimated bias of the model.}

\item{sigma2_hat}{(numeric) The estimated noise of the fixed effects.}

\item{tau_hat}{(numeric) The estimated noise of the random effects.}

\item{K_gpr}{(matrix, n*n) Estimated ensemble kernel matrix.}
}
\value{
\item{test_stat}{(numeric) The computed test statistic.}
}
\description{
Compute score test statistics.
}
\details{
The test statistic is distributed as a scaled Chi-squared distribution.
}
\examples{


compute_stat(n = 100, Y, X12, beta0, sigma2_hat, tau_hat, K_gpr)


}
\references{
Arnab Maity and Xihong Lin. Powerful tests for detecting a gene
effect in the presence of possible gene-gene interactions using garrote
kernel machines. December 2011.
}
\author{
Wenying Deng
}
