% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{ensemble_exp}
\alias{ensemble_exp}
\title{Estimating Ensemble Kernel Matrices Using EXP}
\usage{
ensemble_exp(n, kern_size, beta, error_mat, A_hat)
}
\arguments{
\item{n}{(integer) A numeric number specifying the number of observations.}

\item{kern_size}{(integer, =K) A numeric number specifying the number of 
kernels in the kernel library.}

\item{beta}{(numeric/character) A numeric value specifying the parameter 
when strategy = "exp". See Details.}

\item{error_mat}{(matrix, n*K) A n\*kern_size matrix indicating errors.}

\item{A_hat}{(list of length K) A list of projection matrices for 
every kernels in the kernel library.}
}
\value{
\item{A_est}{(matrix, n*n) A list of estimated kernel matrices.}

\item{u_hat}{(vector of length K) A vector of weights of the kernels 
in the library.}
}
\description{
Give a list of estimated kernel matrices and their weights using
exponential weighting.
}
\details{
\bold{Exponential Weighting}

Additionally, another scholar gives a new strategy to calculate weights
based on the estimated errors \eqn{\{\hat{\epsilon}_d\}_{d=1}^D}.
\deqn{u_d(\beta)=\frac{exp(-\parallel \hat{\epsilon}_d
\parallel_2^2/\beta)}{\sum_{d=1}^Dexp(-\parallel \hat{\epsilon}_d
\parallel_2^2/\beta)}}

\bold{beta}

The value of beta can be "min"=\eqn{min\{RSS\}_{d=1}^D/10},
"med"=\eqn{median\{RSS\}_{d=1}^D}, "max"=\eqn{max\{RSS\}_{d=1}^D*2}
and any other positive numeric number, where \eqn{\{RSS\} _{d=1}^D}
are the set of residual sum of squares of \eqn{D} base kernels.
}
\references{
Jeremiah Zhe Liu and Brent Coull. Robust Hypothesis Test for
Nonlinear Effect with Gaus- sian Processes. October 2017.

Xiang Zhan, Anna Plantinga, Ni Zhao, and Michael C. Wu. A fast small-sample
kernel inde- pendence test for microbiome community-level association
analysis. December 2017.

Arnak S. Dalalyan and Alexandre B. Tsybakov. Aggregation by Exponential
Weighting and Sharp Oracle Inequalities. In Learning Theory, Lecture Notes
in Computer Science, pages 97– 111. Springer, Berlin, Heidelberg, June 2007.
}
\seealso{
mode: \code{\link{tuning}}
}
\author{
Wenying Deng
}
