% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{estimate_base}
\alias{estimate_base}
\title{Estimating Projection Matrices}
\usage{
estimate_base(n, kern_size, Y, X1, X2, kern_list, mode, lambda)
}
\arguments{
\item{n}{(integer) A numeric number specifying the number of observations.}

\item{kern_size}{(integer, =K) A numeric number specifying the number of 
kernels in the kernel library.}

\item{Y}{(vector of length n) Reponses of the dataframe.}

\item{X1}{(dataframe, n*p1) The first type of factor in the dataframe (could 
contains several subfactors).}

\item{X2}{(dataframe, n*p2) The second type of factor in the dataframe (could 
contains several subfactors).}

\item{kern_list}{(list of length K) A list of kernel functions given by user.}

\item{mode}{(character) A character string indicating which tuning parameter 
criteria is to be used.}

\item{lambda}{(numeric) A numeric string specifying the range of noise 
to be chosen. The lower limit of lambda must be above 0.}
}
\value{
\item{A_hat}{(list of length K) A list of projection matrices for 
every kernels in the kernel library.}

\item{error_mat}{(matrix, n*K) A n\*kern_size matrix indicating errors.}
}
\description{
Calculate the estiamted projection matrices for every kernels in the kernel
library.
}
\details{
For a given mode, this function return a list of projection matrices for
every kernels in the kernel library and a n*kern_size matrix indicating
errors.
}
\examples{


estimate_base(n = 100, kern_size = 3, Y, X1, X2, kern_list,
mode = "loocv", lambda = exp(seq(-5, 5)))


}
\references{
Jeremiah Zhe Liu and Brent Coull. Robust Hypothesis Test for
Nonlinear Effect with Gaus- sian Processes. October 2017.
}
\author{
Wenying Deng
}
