% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{estimation}
\alias{estimation}
\title{Conducting Gaussian Process Regression}
\usage{
estimation(Y, X, K_list = list(NULL), mode = "loocv",
  strategy = "stack", beta_exp = 1, lambda = exp(seq(-10, 5)), ...)
}
\arguments{
\item{Y}{(vector of length n) Reponses of the dataframe.}

\item{X}{(dataframe, n*p) Fixed effects variables in the dataframe (could
contains several subfactors).}

\item{K_list}{(list of matrices) A nested list of kernel term matrices. 
The first level corresponds to each base kernel function in kern_func_list, 
the second level corresponds to each kernel term specified in the formula.}

\item{mode}{(character) A character string indicating which tuning parameter
criteria is to be used.}

\item{strategy}{(character) A character string indicating which ensemble
strategy is to be used.}

\item{beta_exp}{(numeric/character) A numeric value specifying the parameter
when strategy = "exp" \code{\link{ensemble_exp}}.}

\item{lambda}{(numeric) A numeric string specifying the range of noise to be
chosen. The lower limit of lambda must be above 0.}

\item{...}{Additional parameters to pass to estimate_ridge.}
}
\value{
\item{lambda}{(numeric) The selected tuning parameter based on the
estimated ensemble kernel matrix.}

\item{beta}{(matrix, p*1) Fixed effects estimator of the model.}

\item{alpha}{(vector of length n) Random effects estimator of the estimated
ensemble kernel matrix.}

\item{K}{(matrix, n*n) Estimated ensemble kernel matrix.}

\item{u_hat}{(vector of length K) A vector of weights of the kernels in the
library.}

\item{base_est}{(list of length 6) The detailed estimation results of K
kernels.}
}
\description{
Conduct gaussian process regression based on the estimated ensemble kernel
matrix.
}
\details{
After obtaining the ensemble kernel matrix, we can calculate the outpur of
gaussian process regression.
}
\examples{

result <- estimation(Y = CVEK:::model_matrices$y, 
X = CVEK:::model_matrices$X, 
K_list = CVEK:::model_matrices$K, 
mode = "loocv", strategy = "stack", 
beta_exp = 1, lambda = exp(seq(-10, 5)))

}
\seealso{
strategy: \code{\link{ensemble}}
}
\author{
Wenying Deng
}
