% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{estimation}
\alias{estimation}
\title{Conducting Gaussian Process Regression}
\usage{
estimation(Y, X1, X2, Kernlist, mode = "loocv", strategy = "erm",
  beta = 1, lambda = exp(seq(-5, 5)))
}
\arguments{
\item{Y}{Reponses of the dataframe.}

\item{X1}{The first type of factor in the dataframe (could contains several
subfactors).}

\item{X2}{The second type of factor in the dataframe (could contains several
subfactors).}

\item{Kernlist}{The kernel library containing several kernels given by user.}

\item{mode}{A character string indicating which tuning parameter criteria is
to be used.}

\item{strategy}{A character string indicating which ensemble strategy is to
be used.}

\item{beta}{A numeric value specifying the parameter when strategy = "exp".}

\item{lambda}{A numeric string specifying the range of noise to be chosen.
The lower limit of lambda must be above 0.}
}
\value{
\item{lam}{The selected tuning parameter based on the estimated
ensemble kernel matrix.}

\item{intercept}{Estimated bias of the model.}

\item{alpha}{Estimated coefficients of the estimated ensemble kernel
matrix.}

\item{K}{Estimated ensemble kernel matrix.}

\item{u_hat}{A vector of weights of the kernels in the library.}
}
\description{
Conduct gaussian process regression based on the estimated ensemble kernel
matrix.
}
\details{
After obtaining the ensemble kernel matrix, we can calculate the outpur of
gaussian process regression, the solution is given by
\deqn{\hat{\beta}=[1^T(K+\lambda I)^{-1}1]^{-1}1^T(K+\lambda I)^{-1}y}
\deqn{\hat{\alpha}=(K+\lambda I)^{-1}(y-\hat{\beta}1)} where
\eqn{\beta=intercept}.
}
\examples{


##estimation(Y, X1, X2, Kernlist, mode = "loocv", strategy = "erm",
##beta = 1, lambda = exp(seq(-5, 5)))


}
\seealso{
strategy: \code{\link{ensemble}}
}
\author{
Wenying Deng
}
