% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{test_asym}
\alias{test_asym}
\title{Conducting Score Tests for Interaction Using Asymptotic Test}
\usage{
test_asym(n, Y, X12, beta0, alpha0, K_gpr, sigma2_hat, tau_hat, B)
}
\arguments{
\item{n}{(integer) A numeric number specifying the number of observations.}

\item{Y}{(vector of length n) Reponses of the dataframe.}

\item{X12}{(dataframe, n*(p1\*p2)) The interaction items of first and second 
types of factors in the dataframe.}

\item{beta0}{(numeric) Estimated bias of the model.}

\item{alpha0}{(vector of length n) Estimated coefficients of the estimated 
ensemble kernel matrix.}

\item{K_gpr}{(matrix, n*n) Estimated ensemble kernel matrix.}

\item{sigma2_hat}{(numeric) The estimated noise of the fixed effects.}

\item{tau_hat}{(numeric) The estimated noise of the random effects.}

\item{B}{(integer) A numeric value indicating times of resampling 
when test = "boot".}
}
\value{
\item{pvalue}{(numeric) p-value of the test.}
}
\description{
Conduct score tests comparing a fitted model and a more general alternative
model using asymptotic test.
}
\details{
\bold{Asymptotic Test}

This is based on the classical variance component test to construct a
testing procedure for the hypothesis about Gaussian process function.
}
\references{
Xihong Lin. Variance component testing in generalised linear
models with random effects. June 1997.

Arnab Maity and Xihong Lin. Powerful tests for detecting a gene effect in
the presence of possible gene-gene interactions using garrote kernel
machines. December 2011.

Petra Bu ̊zˇkova ́, Thomas Lumley, and Kenneth Rice. Permutation and
parametric bootstrap tests for gene-gene and gene-environment interactions.
January 2011.
}
\seealso{
method: \code{\link{generate_kernel}}

mode: \code{\link{tuning}}

strategy: \code{\link{ensemble}}
}
\author{
Wenying Deng
}
